+++ test/type_inference/init.lua
@@ -3652,17 +3652,17 @@ end
 local <?n?> = test(1)
 ]]
 
---TEST 'boolean' [[
------@type boolean, number
---local <?x?>, y
---]]
---
---TEST 'number' [[
------@type boolean, number
---local x, <?y?>
---]]
---
---TEST 'unknown' [[
------@type _, number
---local <?x?>, y
---]]
+TEST 'boolean' [[
+---@type boolean, number
+local <?x?>, y
+]]
+
+TEST 'number' [[
+---@type boolean, number
+local x, <?y?>
+]]
+
+TEST 'unknown' [[
+---@type _, number
+local <?x?>, y
+]]